+++ gql_websocket_link/lib/src/exceptions.dart
@@ -1,0 +1,4 @@
+// Originally from https://github.com/zino-app/graphql-flutter
+// Adapted to `gql` by @iscriptology
+
+// no exceptions yet for the beta period