+++ multipack/lib/commands/fmt.dart
@@ -1,0 +1,19 @@
+import "dart:async";
+
+import "package:multipack/commands/common.dart";
+import "package:multipack/package.dart";
+
+class FmtCommand extends PassthroughCommand {
+  FmtCommand(List<Package> packages)
+      : super(
+          "fmt",
+          "Run formatter.",
+          packages,
+        );
+
+  @override
+  FutureOr<int> runOnPackage(Package package) => package.fmt(
+        argResults.arguments,
+        nameWidth: nameWidth,
+      );
+}